Update faq.md
@bramcohen stated "Probably the vernacular should be that 'mining' refers to all PoW and 'farming' is a more specific term referring just to PoST. Otherwise the discussion is too confusing". Therefore I have replaced the words mining with farming. Added a user question about decentralization.

+ <b>Q - "How is chia resistant to miner centralization?"

A: Chia Network uses vacant hard disk drive (HDD) space to farm seeds. As there is so much HDD space available and anyone with HDD space will be able to farm, this makes farming decentralized. A Bitcoin Asic centralized mining system makes the formation of blocks a centralized process because ASIC mining and electricity costs are beyond many people's affordability. Manufacturers of HDD could optimize to farm faster, but the cost of doing so is not worth the reward. Thus Chia will be the most decentralized cryptocurrency on the market.