Important word change
The word "which" is used to include incidental information about a sentence, i.e. information that does not change its meaning. With Chia, we are not creating a cryptocurrency that, oh, by the way, just happens to be less wasteful, more decentralized, and more secure - rather, these attributes are vital characteristics of the cryptocurrency in question. It is far more appropriate to use "that" in the sentence serving as the website's front-and-center description of the project's mission.